requirements.txt
Django==5.0.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-environ==0.9.0
